Move RtcConfiguration into separate module
