+++ crates/native/src/api/rtc_configuration/rtc_ice_server.rs
@@ -1,0 +1,24 @@
+//! Description of STUN and TURN servers.
+
+/// Description of STUN and TURN servers that can be used by an [ICE Agent][1]
+/// to establish a connection with a peer.
+///
+/// [1]: https://w3.org/TR/webrtc#dfn-ice-agent
+#[derive(Debug)]
+pub struct RtcIceServer {
+    /// STUN or TURN URI(s).
+    pub urls: Vec<String>,
+
+    /// If this [`RtcIceServer`] object represents a TURN server, then this
+    /// attribute specifies the [username][1] to use with that TURN server.
+    ///
+    /// [1]: https://w3.org/TR/webrtc#dom-rtciceserver-username
+    pub username: String,
+
+    /// If this [`RtcIceServer`] object represents a TURN server, then this
+    /// attribute specifies the [credential][1] to use with that TURN
+    /// server.
+    ///
+    /// [1]: https://w3.org/TR/webrtc#dom-rtciceserver-credential
+    pub credential: String,
+}
