+++ lib/src/api/bridge/frb.dart
@@ -15,6 +15,10 @@ export 'api/peer_connection_event/ice_gathering_state.dart';
 export 'api/peer_connection_event/peer_connection_state.dart';
 export 'api/peer_connection_event/rtc_track_event.dart';
 export 'api/peer_connection_event/signaling_state.dart';
+export 'api/rtc_configuration.dart';
+export 'api/rtc_configuration/bundle_policy.dart';
+export 'api/rtc_configuration/ice_transports_type.dart';
+export 'api/rtc_configuration/rtc_ice_server.dart';
 export 'api/rtc_ice_candidate_stats.dart';
 export 'api/rtc_rtp_encoding_parameters.dart';
 export 'api/rtc_rtp_send_parameters.dart';
